* arts-2-1-701
* src/tmatrix.{cc,h}: Add checks for ref index size.

* src/xml_io_compound_types.cc: Fix typo.

* src/wigner_functions.cc (wigner3j):

Add braces for outer if to avoid confusion where the inner else belongs
to.

(wigner9j): Return NAN to avoid warning.

* src/test_matpack (test45): Use explicit cast to avoid warning.

git-svn-id: https://www.sat.ltu.se/svn/rt/arts/trunk@8544 aaf1aab0-4228-0410-ad68-8dceda47f409

+++ src/test_matpack.cc
@@ -940,7 +940,7 @@ ArrayOfIndex J;J=MakeArray<Index>(0, 2, 2, 4, 4, 6, 6, 8, 8, 10, 10, 12, 12, 14,
     Vector d; d.resize(38);
     
     for(Index II = 0; II<N.nelem(); II++)
-        d[II] = wigner6j(1, 1, 1, J[II], N[II], N[II]) * pow(-1.,2.*N[II]) * sqrt(6*(2*N[II]+1)*(2*J[II]+1));
+        d[II] = wigner6j(1, 1, 1, J[II], N[II], N[II]) * pow(-1.,2.*(Numeric)N[II]) * sqrt(6*(2*N[II]+1)*(2*J[II]+1));
     
     std::cout<<d<<std::endl;
